+++ front/src/App.tsx
@@ -6,6 +6,8 @@ import BaseLayout from './components/Layouts/BaseLayout';
 import MainPage from './pages/MainPage';
 import MyProfilePage from './pages/MyProfilePage';
 import LetterListPage from './pages/LetterListPage';
+import LoginPage from './pages/LoginPage';
+import SignUpPage from './pages/SignUpPage';
 
 function App() {
   return (
@@ -15,8 +17,8 @@ function App() {
           <Routes>
             <Route element={<NoneLayout />}>
               <Route path="/" element={<Intro />} />
-              <Route path="/login" element={<Intro />} />
-              <Route path="/signup" element={<Intro />} />
+              <Route path="/login" element={<LoginPage />} />
+              <Route path="/signup" element={<SignUpPage />} />
             </Route>
             <Route element={<BaseLayout isAuth />}>
               <Route path="/main" element={<MainPage />} />
