+++ front/src/components/Login/LoginMain.module.scss
@@ -1,0 +1,46 @@
+.login_container {
+  background-color: var(--white);
+  padding: 3rem 2rem;
+  color: var(--button-primary-color);
+  font-size: 0.75rem;
+  font-weight: bold;
+  margin-bottom: 1.875rem;
+  box-shadow: 0px 0px 30px rgba(0, 0, 0, 0.1);
+  .inputform_lable {
+    display: block;
+    width: 100%;
+    margin-bottom: 0.3125rem;
+  }
+  .inputform_input {
+    height: 1.875rem;
+    width: 100%;
+    border: none;
+    border-bottom: 0.125rem solid black;
+    margin-bottom: 2rem;
+    padding: 0.125rem;
+    padding-left: 1.5625rem;
+    border-color: var(--text-primary-color);
+    &:focus {
+      border: none;
+      outline-color: var(--text-primary-color);
+    }
+  }
+}
+.icon {
+  position: absolute;
+  top: 1.5rem;
+  left: 0.1875rem;
+  color: var(--text-primary-color);
+  width: 1.125rem;
+  height: 1.125rem;
+}
+.signUp_Btn {
+  margin-top: 0.3125rem;
+  width: 100%;
+  color: white;
+  font-size: 1.25rem;
+  padding: 0.9375rem;
+  background-image: url(../../assets/DarkBtn.svg);
+  background-repeat: no-repeat;
+  background-size: cover;
+}
