+++ front/src/components/SignUp/SignUp.module.scss
@@ -1,6 +1,6 @@
 .container {
+  background-color: var(--white);
   padding: 2rem;
-  width: 37.5rem;
   color: var(--button-primary-color);
   font-size: 0.75rem;
   font-weight: bold;
@@ -8,34 +8,6 @@
   box-shadow: 0px 0px 30px rgba(0, 0, 0, 0.1);
   p {
     margin-bottom: 0.3125rem;
-  }
-}
-.login_container {
-  padding: 3rem 2rem;
-  width: 37.5rem;
-  color: var(--button-primary-color);
-  font-size: 0.75rem;
-  font-weight: bold;
-  margin-bottom: 1.875rem;
-  box-shadow: 0px 0px 30px rgba(0, 0, 0, 0.1);
-  .inputform_lable {
-    display: block;
-    width: 100%;
-    margin-bottom: 0.3125rem;
-  }
-  .inputform_input {
-    height: 1.875rem;
-    width: 100%;
-    border: none;
-    border-bottom: 0.125rem solid black;
-    margin-bottom: 2rem;
-    padding: 0.125rem;
-    padding-left: 1.5625rem;
-    border-color: var(--text-primary-color);
-    &:focus {
-      border: none;
-      outline-color: var(--text-primary-color);
-    }
   }
 }
 
@@ -126,14 +98,7 @@
   color: white;
   font-size: 1.25rem;
   padding: 0.9375rem;
-  background-image: url(./Btn.svg);
+  background-image: url(../../assets/DarkBtn.svg);
   background-repeat: no-repeat;
   background-size: cover;
-}
-
-body {
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
 }
