+++ front/src/components/SignUp/SignUpFooter.module.scss
@@ -1,0 +1,12 @@
+.signup_footer {
+  width: 100%;
+  text-align: center;
+  font-size: 0.75rem;
+  P {
+    color: #666666;
+    margin-bottom: 0.25rem;
+  }
+  a {
+    color: var(--button-primary-color);
+  }
+}
